Bug 33705: (bug 33066 follow-up) Fix 'configure' button for kohaTable
The 'configure' button is displayed for any tables, even those that
cannot be configured.

This is a regression caused by
  commit 765fd1ced3b9efc4ff6fb71e2fee1a7a227d1cae
  Bug 33066: Introduce a KohaTable Vue component

It's adding a default value for table_settings, but then later we are
testing if table_settings is true:
699     let table_settings = params.table_settings || {};
...
798     if ( table_settings && CAN_user_parameters_manage_column_config ) {

This patch is reverting the default value, so the test will be
corrected.

Test plan:
1. Go to Tools > Quote editor
2. Click "Configure"
=> Without this patch you are brought to the column settings page, but the quote editor table is not there
=> With this patch the button is not present
3. Go to the cities page (admin/cities.pl)
4. Click "Configure"
=> You are brought to the column settings page

Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>
Signed-off-by: Victor Grousset/tuxayo <victor@tuxayo.net>