Bug 33892: Use template wrapper for tabs: OPAC authority detail
This patch updates the OPAC authority detail page so that it uses
the new WRAPPER syntax to generate tabs markup.

To test, apply the patch and locate an authority record in the OPAC
which has notes fields.

When viewing the details for that record the notes should appear in a
single "Notes" tab.

Signed-off-by: Sam Lau <samalau@gmail.com>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit afd30b8f4d11c36c86f26a8d5849c7990297f265)
Signed-off-by: Matt Blenkinsop <matt.blenkinsop@ptfs-europe.com>