Bug 33921: Improve translation of title tags: Plugins and Point of sale
This patch updates plugin and point of sale templates so
that title tags can be more easily translated.

The patch also makes some minor changes to some templates to improve
consistency between page title, breadcrumb navigation, and page heading.

To test, apply the patch and confirm that the following pages have the
correct title tags:

- Administration ->
  - Manage plugins
  - Upload plugin

  * With 'useplugins' set to 0 in koha-conf.xml you can navigate
  directly to /cgi-bin/koha/plugins/plugins-home.pl to confirm that
  the "Plugins disabled" page is shown.

With EnablePointOfSale and UseCashRegisters enabled:

- Point of sale
  - Make a sale and print a receipt for the transaction
  - Transaction history for [register]
  - Cash summary for [library]

Signed-off-by: Andrew Auld <andrew.auld@ptfs-europe.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>