Bug 33924: Improve translation of title tags: Rotating collections
This patch updates rotating collection templates so that title tags can
be more easily translated.

The patch also makes some corrections to breadcrumb navigation and page
headers to make sure they are consistent with titles.

To test, apply the patch and confirm that the following rotating
collection pages have the correct title tags:

- Tools -> Rotating collections
  - Rotating collections list
  -> New collection
  -> Edit collection
  -> Delete collection -> Confirmation page
  -> Transfer collection

Signed-off-by: Salah Ghedda <salah.ghedda@inLibro.com>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>