Bug 33951: Set normalized_oclc in opac/opac-readingrecord.pl
normalized_oclc is used in the template to build the link to syndetics,
but it's not passed from the controller. Is this patch correct? Is
syndetics broken on this page? Should we remove it or keep it?
Is oclc parameter in the URL unecessary and should be removed?

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 8e93668796fd579e4dcbbbdf01b316ed044bd5bc)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
(cherry picked from commit 9f0a6d53df111ed746a5b558a1867f42c9f2189c)
Signed-off-by: Matt Blenkinsop <matt.blenkinsop@ptfs-europe.com>